Made the fix a filter of it's own for better clarity. Added regression test.
