+++ src/javasources/KTool/src/org/kframework/parser/DefinitionLoader.java
@@ -28,6 +28,7 @@ import org.kframework.parser.concrete.disambiguate.BestFitFilter;
 import org.kframework.parser.concrete.disambiguate.CellEndLabelFilter;
 import org.kframework.parser.concrete.disambiguate.CellTypesFilter;
 import org.kframework.parser.concrete.disambiguate.CorrectCastPriorityFilter;
+import org.kframework.parser.concrete.disambiguate.CorrectConstantsTransformer;
 import org.kframework.parser.concrete.disambiguate.CorrectKSeqFilter;
 import org.kframework.parser.concrete.disambiguate.CorrectRewritePriorityFilter;
 import org.kframework.parser.concrete.disambiguate.FlattenListsFilter;
@@ -85,7 +86,7 @@ public class DefinitionLoader {
 
     /**
      * step. 1. slurp 2. gen files 3. gen TBLs 4. import files in stratego 5. parse configs 6. parse rules 7. ???
-     * 
+     *
      * @param mainFile
      * @param mainModule
      * @return
@@ -226,7 +227,7 @@ public class DefinitionLoader {
 
     /**
      * Parses a string representing a file with modules in it. Returns the complete parse tree. Any bubble rule has been parsed and disambiguated.
-     * 
+     *
      * @param content
      *            - the input string.
      * @param filename
@@ -304,6 +305,7 @@ public class DefinitionLoader {
         // config = config.accept(new TypeInferenceSupremumFilter(context));
         config = config.accept(new BestFitFilter(new GetFitnessUnitKCheckVisitor(context), context));
         config = config.accept(new PreferAvoidFilter(context));
+        config = config.accept(new CorrectConstantsTransformer(context));
         config = config.accept(new FlattenListsFilter(context));
         config = config.accept(new AmbDuplicateFilter(context));
         // last resort disambiguation
@@ -357,6 +359,7 @@ public class DefinitionLoader {
         // config = config.accept(new TypeInferenceSupremumFilter(context));
         config = config.accept(new BestFitFilter(new GetFitnessUnitKCheckVisitor(context), context));
         config = config.accept(new PreferAvoidFilter(context));
+        config = config.accept(new CorrectConstantsTransformer(context));
         config = config.accept(new FlattenListsFilter(context));
         config = config.accept(new AmbDuplicateFilter(context));
         // last resort disambiguation
