+++ src/javasources/KTool/src/org/kframework/parser/generator/ParseRulesFilter.java
@@ -20,6 +20,7 @@ import org.kframework.parser.concrete.disambiguate.BestFitFilter;
 import org.kframework.parser.concrete.disambiguate.CellEndLabelFilter;
 import org.kframework.parser.concrete.disambiguate.CellTypesFilter;
 import org.kframework.parser.concrete.disambiguate.CorrectCastPriorityFilter;
+import org.kframework.parser.concrete.disambiguate.CorrectConstantsTransformer;
 import org.kframework.parser.concrete.disambiguate.CorrectKSeqFilter;
 import org.kframework.parser.concrete.disambiguate.CorrectRewritePriorityFilter;
 import org.kframework.parser.concrete.disambiguate.FlattenListsFilter;
@@ -161,6 +162,7 @@ public class ParseRulesFilter extends BasicTransformer {
                 // config = config.accept(new TypeInferenceSupremumFilter(context));
                 config = config.accept(new BestFitFilter(new GetFitnessUnitKCheckVisitor(context), context));
                 config = config.accept(new PreferAvoidFilter(context));
+                config = config.accept(new CorrectConstantsTransformer(context));
                 config = config.accept(new FlattenListsFilter(context));
                 config = config.accept(new AmbDuplicateFilter(context));
                 // last resort disambiguation
