+++ tests/config.xml
@@ -381,6 +381,27 @@
   </test>
 
   <test
+      definition="tests/regression/tokenconstants/issue.k"
+      programs="tests/regression/tokenconstants/programs"
+      results="tests/regression/tokenconstants/out"
+      extension="issue">
+    <program name="test.issue">
+      <krun-option name="-color" value="off" />
+    </program>
+  </test>
+
+  <test
+      definition="tests/regression/tokenconstants/issue.k"
+      programs="tests/regression/tokenconstants/programs"
+      results="tests/regression/tokenconstants/out"
+      extension="issue">
+    <kompile-option name="-backend" value="java" />
+    <program name="test.issue">
+      <krun-option name="-color" value="off" />
+    </program>
+  </test>
+  
+  <test
       definition="tests/regression/unparsing/issue.k"
       programs="tests/regression/unparsing/programs"
       extension="issue"
