+++ tests/regression/tokenconstants/test.issue
@@ -1,0 +1,1 @@
+Main
