+++ tests/regression/tokenconstants/test.issue.out
@@ -1,0 +1,3 @@
+<k>
+    .K
+</k>
